+++ src/main/java/Yapper/YapperBot.java
@@ -1,16 +1,6 @@
-package base;
+package Yapper;
 import commands.Command;
 import tasks.*;
-
-import java.io.FileWriter;
-import java.io.IOException;
-import java.time.LocalDateTime;
-import java.time.format.DateTimeFormatter;
-import java.time.format.DateTimeParseException;
-import java.util.Objects;
-import java.util.*;
-import java.util.function.Consumer;
-import java.util.regex.*;
 
 
 public class YapperBot {
