+++ src/main/java/commands/Command.java
@@ -1,7 +1,7 @@
 package commands;
 
-import base.Storage;
-import base.Ui;
+import Yapper.Storage;
+import Yapper.Ui;
 import tasks.TaskList;
 
 public interface Command {
