+++ src/main/java/commands/DeadlineCommand.java
@@ -1,7 +1,7 @@
 package commands;
 
-import base.Storage;
-import base.Ui;
+import Yapper.Storage;
+import Yapper.Ui;
 import tasks.Deadline;
 import tasks.Task;
 import tasks.TaskList;
