+++ src/main/java/commands/DeleteCommand.java
@@ -1,8 +1,7 @@
 package commands;
 
-import base.InvalidInputException;
-import base.Storage;
-import base.Ui;
+import Yapper.Storage;
+import Yapper.Ui;
 import tasks.Task;
 import tasks.TaskList;
