+++ src/main/java/commands/EventCommand.java
@@ -1,8 +1,7 @@
 package commands;
 
-import base.Storage;
-import base.Ui;
-import tasks.Deadline;
+import Yapper.Storage;
+import Yapper.Ui;
 import tasks.Event;
 import tasks.Task;
 import tasks.TaskList;
